# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "IM Chat backend API documentation.",
        "title": "IM Chat API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/chat/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "查看历史聊天记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "对方用户 ID",
                        "name": "receiver_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认 20，最大 100",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HistoryResponse"
                        }
                    }
                }
            }
        },
        "/chat/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "更改消息为已读",
                "parameters": [
                    {
                        "description": "会话已读请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReadContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/chat/show/all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "展示所有未被删除的聊天框（侧边框）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contact"
                        }
                    }
                }
            }
        },
...
        "/user/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "完善用户资料",
                "parameters": [
                    {
                        "description": "完善资料：name 空则保留原名；gender/signature/picture_id 省略则不改；signature 可传空串清空；picture_id 为图片库 id（≥1），\u003c1 时按默认头像 id=1 保存",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserMain2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/deleteUser": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/heartbeat": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "心跳续期",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    }
                }
            }
        },
        "/user/pwd/code/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "发送验证码以便验证身份",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterServer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/show/main": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "显示当前用户资料",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserMainReturn"
                        }
                    }
                }
            }
        },
        "/user/show/pictures": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 picture 表全部 id、web（需登录，无请求体）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "展示图片库",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PictureMainReturn"
                        }
                    },
                    "401": {
                        "description": "未登录或 token 无效",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/update/email": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "修改邮箱",
                "parameters": [
                    {
                        "description": "修改邮箱请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/update/pwd": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ReUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ChatMsg": {
            "type": "object",
            "required": [
                "context",
                "msg_type",
                "receiver_id"
            ],
            "properties": {
                "context": {
                    "type": "string"
                },
                "msg_type": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "integer"
                }
            }
        },
        "models.Contact": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ListContact"
                    }
                }
            }
        },
...
        "models.HistoryMsg": {
            "type": "object",
            "required": [
                "receiver_id"
            ],
            "properties": {
                "msg": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChatMsg"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "receiver_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "models.HistoryResponse": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "msg": {
                    "$ref": "#/definitions/models.HistoryMsg"
                }
            }
        },
...
        "models.ListContact": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "friend_name": {
                    "type": "string"
                },
                "friend_picture": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.LoginResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.LoginServer": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.PictureMainReturn": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "picture": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Pictures"
                    }
                }
            }
        },
        "models.Pictures": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "web": {
                    "type": "string"
                }
            }
        },
        "models.ReEmail": {
            "type": "object",
            "required": [
                "code",
                "new_email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "new_email": {
                    "type": "string"
                }
            }
        },
        "models.ReUpdate": {
            "type": "object",
            "required": [
                "code",
                "new_password"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
        "models.ReadContact": {
            "type": "object",
            "required": [
                "friend_id"
            ],
            "properties": {
                "friend_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.RegisterServer": {
            "type": "object",
            "required": [
                "code",
                "email",
                "password",
                "re_password"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "re_password": {
                    "type": "string"
                }
            }
        },
        "models.UserMain": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "description": "0为男 1为女 2为未知",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "picture_id": {
                    "type": "integer"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "models.UserMain2": {
            "type": "object",
            "properties": {
                "gender": {
                    "description": "0男 1女 2未知",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "picture_id": {
                    "type": "integer"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "models.UserMainReturn": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.UserMain"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}